# swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a simple example NodeJS API project to demonstrate Swagger Documentation",
        "version": "1.0.0",
        "title": "RefugeeHelp API",
        "contact": {
            "email": "abc@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "schemes": ["http"],
    "host": "localhost:5000",
    "basePath": "/",
    "paths" : {
        "/points" : {
            "get" : {
                "summary" : "Get all points in DB",
                "description" : "Get all public points available for all users",
                "produces" : ["application/json"],
                "responses" : {
                    "200" : {
                        "description" : "Success"
                    }
                }
            },
            "post" : {
                "summary" : "Add new point",
                "description" : "Add new point to public points DB",
                "produces" : ["application/json"],
                "parameters" : [
                    {
                        "in" : "body",
                        "name" : "User",
                        "description" : "Point Object",
                        "required" : true,
                        "type" : "object"
                    }
                ],
                "responses" : {
                    "201" : {
                        "description" : "Success"
                    },
                    "400" : {
                        "description" : "Bad request"
                    },
                    "403" : {
                        "description" : "Forbidden"
                    },
                    "500" : {
                        "description" : "Internal server error"
                    }
                }
            },
            "delete" : {
                "summary" : "Delete the point with the given name",
                "description" : "Delete the point with the given name",
                "produces" : ["application/json"],
                "parameters" : [
                    {
                        "in" : "body",
                        "name" : "User",
                        "description" : "Point Object",
                        "required" : true,
                        "type" : "object"
                    }
                ],
                "responses" : {
                    "200" : {
                        "description" : "Success"
                    },
                    "400" : {
                        "description" : "Bad request"
                    },
                    "403" : {
                        "description" : "Forbidden"
                    },
                    "404" : {
                        "description" : "Not found"
                    },
                    "500" : {
                        "description" : "Internal server error"
                    }
                }
            },
            "put" : {
                "summary" : "Update the point with the given name",
                "description" : "Update the point description with the given name",
                "produces" : ["application/json"],
                "parameters" : [
                    {
                        "in" : "body",
                        "name" : "User",
                        "description" : "Point Object",
                        "required" : true,
                        "type" : "object"
                    }
                ],
                "responses" : {
                    "200" : {
                        "description" : "Success"
                    },
                    "400" : {
                        "description" : "Bad request"
                    },
                    "403" : {
                        "description" : "Forbidden"
                    },
                    "404" : {
                        "description" : "Not found"
                    },
                    "500" : {
                        "description" : "Internal server error"
                    }
                }
            }
        },
        "/users" : {
            "get" : {
                "summary" : "Get all users in DB",
                "description" : "Returns a list with all users",
                "produces" : ["application/json"],
                "responses" : {
                    "200" : {
                        "description" : "Success"
                    }
                }
            },
            "post" : {
                "summary" : "Add new users",
                "description" : "Add new oint to public points DB",
                "produces" : ["application/json"],
                "parameters" : [
                    {
                        "in" : "body",
                        "name" : "User",
                        "description" : "User Object",
                        "required" : true,
                        "type" : "object"
                    }
                ],
                "responses" : {
                    "201" : {
                        "description" : "Success"
                    },
                    "400" : {
                        "description" : "Bad request"
                    },
                    "403" : {
                        "description" : "Forbidden"
                    },
                    "500" : {
                        "description" : "Internal server error"
                    }
                }
            },
            "delete" : {
                "summary" : "Delete the user with the given email",
                "description" : "Delete the user with the email",
                "produces" : ["application/json"],
                "parameters" : [
                    {
                        "in" : "body",
                        "name" : "User",
                        "description" : "User Object",
                        "required" : true,
                        "type" : "object"
                    }
                ],
                "responses" : {
                    "200" : {
                        "description" : "Success"
                    },
                    "400" : {
                        "description" : "Bad request"
                    },
                    "403" : {
                        "description" : "Forbidden"
                    },
                    "404" : {
                        "description" : "Not found"
                    },
                    "500" : {
                        "description" : "Internal server error"
                    }
                }
            },
            "put" : {
                "summary" : "Update the user with the given email",
                "description" : "Update the users description with the given email",
                "produces" : ["application/json"],
                "parameters" : [
                    {
                        "in" : "body",
                        "name" : "User",
                        "description" : "User Object",
                        "required" : true,
                        "type" : "object"
                    }
                ],
                "responses" : {
                    "200" : {
                        "description" : "Success"
                    },
                    "400" : {
                        "description" : "Bad request"
                    },
                    "403" : {
                        "description" : "Forbidden"
                    },
                    "404" : {
                        "description" : "Not found"
                    },
                    "500" : {
                        "description" : "Internal server error"
                    }
                }
            }
        },
        "/users/points" : {
            "post" : {
                "summary" : "Push the object to the user mail",
                "description" : "Add point to user",
                "produces" : ["application/json"],
                "parameters" : [
                    {
                        "in" : "body",
                        "name" : "User",
                        "description" : "User Object",
                        "required" : true,
                        "type" : "object"
                    }
                ],
                "responses" : {
                    "200" : {
                        "description" : "Success"
                    },
                    "400" : {
                        "description" : "Bad request"
                    },
                    "403" : {
                        "description" : "Forbidden"
                    },
                    "404" : {
                        "description" : "Not found"
                    },
                    "500" : {
                        "description" : "Internal server error"
                    }
                }
            },
            "delete" : {
                "summary" : "Remove the point from the user mail",
                "description" : "Remove point from user",
                "produces" : ["application/json"],
                "parameters" : [
                    {
                        "in" : "body",
                        "name" : "User",
                        "description" : "User Object",
                        "required" : true,
                        "type" : "object"
                    }
                ],
                "responses" : {
                    "200" : {
                        "description" : "Success"
                    },
                    "400" : {
                        "description" : "Bad request"
                    },
                    "403" : {
                        "description" : "Forbidden"
                    },
                    "404" : {
                        "description" : "Not found"
                    },
                    "500" : {
                        "description" : "Internal server error"
                    }
                }
            }
        }
    }
}